docintel/pyproject.toml
[project]
name = "docintel"
version = "0.1.0"
description = "Production-ready document intelligence platform with RAG capabilities"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.2",
    "anthropic>=0.8.0",
    "openai>=1.3.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "markdown>=3.5",
    "tiktoken>=0.5.1",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.5",
    "mypy>=1.7.0",
]

[project.scripts]
docintel = "docintel.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
